LayoutTests/imported/w3c/web-platform-tests/trusted-types/trusted-types-source-file-path-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'trusted-types'.

CONSOLE MESSAGE: ReferenceError: Can't find variable: trustedTypes

Harness Error (FAIL), message = ReferenceError: Can't find variable: trustedTypes
